Desktop new 0.1.8.9
A foglaláskezelés átalakítása és az adminisztrációs felhasználói felület fejlesztése

- A foglalások törlési logikáját átalakítottuk a megbízhatóság és az áttekinthetőség érdekében; a ComboBox mostantól közvetlenül a foglalási objektumokhoz kapcsolódik, és azonosítókat használ.
- Frissítettük a GetFoglalasIdByName függvényt, hogy a tényleges foglalási azonosítót vagy -1 értéket adjon vissza.
- Az egységesség érdekében a foglalásokhoz kapcsolódó azonosítókat long típusról int típusra változtattuk.
- Javítottuk a bejelentkezési folyamatot: korlátoztuk a hozzáférést az adminisztrátorokra, hozzáadtunk hiba kezelést és javítottuk az üdvözlő üzeneteket.
- A felhasználó törlése most frissíti a felhasználói listát és visszajelzést ad.
- Frissítettük az oldalsáv színátmenetét és növeltük az ablak minimális szélességét a jobb felhasználói felület érdekében.

+++ Desktop/CyberNest_Admin/CyberNest_Admin/Foglalas.cs
@@ -22,16 +22,16 @@ namespace CyberNest_Admin
     public partial class Foglalas
     {
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonPropertyName("felhasznalo_id")]
-        public long FelhasznaloId { get; set; }
+        public int FelhasznaloId { get; set; }
 
         [JsonPropertyName("eszkoz_id")]
-        public long EszkozId { get; set; }
+        public int EszkozId { get; set; }
 
         [JsonPropertyName("idopont_id")]
-        public long IdopontId { get; set; }
+        public int IdopontId { get; set; }
 
         [JsonPropertyName("berlesi_kezdete")]
         public DateTimeOffset BerlesiKezdete { get; set; }
